ADD : EM parameters updated
`SetDeexcitationIgnoreCut` was set to true for version 11 instead of false for version 10.02. Put it explicitly false on the physic list solved the issue of the factor 10 for the computation time. 

Some comments about this option : 

In Geant4 version 11, the `SetDeexcitationIgnoreCut` method is used to control the behavior of the `G4RadioactiveDecay`process during the simulation. This method sets a flag to ignore the cut imposed by the user on the production of secondary particles by radioactive decays.

By default, when a radioactive decay occurs and produces one or more secondary particles (e.g. electrons, positrons, photons), these particles are subject to the same range cut and production thresholds as any other particle in the simulation. This means that if the range of a secondary particle is smaller than the range cut or the production threshold, it may be discarded by the simulation.

However, in some cases it may be desirable to ignore these cuts for secondary particles produced by radioactive decays. This can be achieved by calling the `SetDeexcitationIgnoreCut` method with a value of true. This will cause the range cut and production thresholds to be ignored for all secondary particles produced by the `G4RadioactiveDecay` process.

It should be noted that ignoring the range cut and production thresholds can lead to longer simulation times and may affect the accuracy of the simulation. Therefore, this feature should be used with caution and only when it is necessary to model the specific physics of the problem being studied.